+++ src/express/routes.ts
@@ -1,8 +1,8 @@
 import { Router } from "express";
-import { populateDB, saveLiveevent, firebaseStatus } from "./endpoints/controls";
-import { getLiveevent, getOGs, getStrikes } from "./endpoints/api";
-import { sharePost } from "./endpoints/share";
-import { webhookTriggered } from "./endpoints/webhook";
+import { populateDB, saveLiveevent, firebaseStatus } from "../endpoints/controls";
+import { getLiveevent, getOGs, getStrikes } from "../endpoints/api";
+import { sharePost } from "../endpoints/share";
+import { webhookTriggered } from "../endpoints/webhook";
 
 export const controlsRoutes = Router();
 export const liveeventRoutes = Router();
